+++ e2e/e2etests/test_stress_eth_withdraw.go
@@ -14,7 +14,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/require"
-	"github.com/zeta-chain/protocol-contracts/pkg/gatewayzevm.sol"
+	"github.com/zeta-chain/protocol-contracts-evm/pkg/gatewayzevm.sol"
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
@@ -107,7 +107,7 @@ func TestStressEtherWithdraw(r *runner.E2ERunner, args []string) {
 				auth.Nonce = big.NewInt(int64(currentNonce))
 
 				// Send withdrawal transaction
-				tx, err := r.GatewayZEVM.Withdraw0(
+				tx, err := r.GatewayZEVM.Withdraw(
 					&auth,
 					r.EVMAddress().Bytes(),
 					withdrawalAmount,