+++ zetaclient/chains/evm/signer/v2_sign.go
@@ -6,10 +6,10 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/pkg/errors"
-	erc20custodyv2 "github.com/zeta-chain/protocol-contracts/pkg/erc20custody.sol"
-	"github.com/zeta-chain/protocol-contracts/pkg/gatewayevm.sol"
-	"github.com/zeta-chain/protocol-contracts/pkg/revert.sol"
-	connector "github.com/zeta-chain/protocol-contracts/pkg/zetaconnectornative.sol"
+	erc20custodyv2 "github.com/zeta-chain/protocol-contracts-evm/pkg/erc20custody.sol"
+	"github.com/zeta-chain/protocol-contracts-evm/pkg/gatewayevm.sol"
+	"github.com/zeta-chain/protocol-contracts-evm/pkg/revert.sol"
+	connector "github.com/zeta-chain/protocol-contracts-evm/pkg/zetaconnectornative.sol"
 )
 
 func must[T any](v T, err error) T {